test: build the worker advisor tests only where worker exists
wing_advisor_test.go carries no build tag, but the tests that drive
worker.observeAdvisor reference worker, which wing_transport.go defines only
for linux and darwin. The package therefore failed to compile for
GOOS=windows, which CI runs.

go build ./... did not catch it — it does not compile test files — so the
cross-platform check quoted in the handoff passed while CI was red. Vetting
catches it, since go vet builds the test package:

    GOOS=windows GOARCH=amd64 go vet ./...

Move the four worker-level tests into wing_advisor_worker_test.go behind
//go:build linux || darwin. The advisor's own threshold tests stay untagged
and keep running on every platform, matching wing_advisor.go, which is itself
platform-neutral.
